move CommonsPostprocessor project agnostic stuff to new file PostprocessForSwift.cs
remove Clash of Streamers specific code
