--- Packages/com.gamingforgood.mobile_build_scripts/Editor/CommonsPostprocessor.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 92f7a5eab9de4786956589fc635f9a34
-timeCreated: 1580379083
