+++ Packages/com.gamingforgood.mobile_build_scripts/Editor/PostprocessForSwift.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 80caf10659fae4f35ab8b4ba1fb15035
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
